+++ mixxx/src/controlobject.h
@@ -42,7 +42,7 @@ struct QueueObjectMidi
 {
     ControlObject *pControlObject;
     MidiCategory category;
-    int value;
+    double value;
 };
 
 /**
@@ -89,7 +89,7 @@ public:
     /** Queue a control change from a widget. Thread safe. Blocking. */
     void queueFromThread(double dValue, ControlObjectThread *pControlObjectThread=0);
     /** Queue a control change from MIDI. Thread safe. Blocking. */
-    void queueFromMidi(MidiCategory c, int v);
+    void queueFromMidi(MidiCategory c, double v);
     /** Return a ControlObject value, corresponding to the widget input value. Thread safe. */
     virtual double getValueFromWidget(double dValue);
     /** Return a widget value corresponding to the ControlObject input value. Thread safe. */
@@ -107,7 +107,7 @@ protected:
     /** Sets the value of the object. Not thread safe. */
     virtual void setValueFromEngine(double dValue);
     /** Called when a widget has changed value. Not thread safe. */
-    virtual void setValueFromMidi(MidiCategory, int v);
+    virtual void setValueFromMidi(MidiCategory, double v);
     /** Called when another thread has changed value. Not thread safe. */
     virtual void setValueFromThread(double dValue);
